move loop folder back
