+++ lib/subscription/createCryptoSubscription.ts
@@ -1,7 +1,7 @@
 import { prisma } from '@charmverse/core/prisma-client';
 
+import { getLoopProducts } from 'lib/loop/loop';
 import { NotFoundError } from 'lib/middleware';
-import { getLoopProducts } from 'lib/subscription/loop/loop';
 
 import { communityProduct, loopCheckoutUrl } from './constants';
 import { getCouponDetails } from './getCouponDetails';
