+++ pages/api/v1/webhooks/stripe/index.ts
@@ -5,11 +5,11 @@ import { prisma } from '@charmverse/core/prisma-client';
 import type { NextApiRequest, NextApiResponse } from 'next';
 import type Stripe from 'stripe';
 
+import { getLoopProducts } from 'lib/loop/loop';
 import { trackUserAction } from 'lib/metrics/mixpanel/trackUserAction';
 import { defaultHandler } from 'lib/public-api/handler';
 import { communityProduct, loopCheckoutUrl } from 'lib/subscription/constants';
 import { getActiveSpaceSubscription } from 'lib/subscription/getActiveSpaceSubscription';
-import { getLoopProducts } from 'lib/subscription/loop/loop';
 import { stripeClient } from 'lib/subscription/stripe';
 import { relay } from 'lib/websockets/relay';
